depreciate pregnancy test and move the goods into patient test
