+++ app/models/patient.rb
@@ -15,6 +15,7 @@ class Patient
   include Notetakeable
   include Searchable
   include AttributeDisplayable
+  include Pledgeable
 
   LINES.each do |line|
     scope line.downcase.to_sym, -> { where(:_line.in => [line]) }
@@ -168,16 +169,5 @@ class Patient
     name.strip! if name
     other_contact.strip! if other_contact
     other_contact_relationship.strip! if other_contact_relationship
-  end
-
-  def updating_pledge_sent?
-    pledge_sent == true
-  end
-
-  def pledge_info_presence
-    errors.add(:pledge_sent, 'DCAF soft pledge field cannot be blank') if dcaf_soft_pledge.blank?
-    errors.add(:pledge_sent, 'Patient name cannot be blank') if name.blank?
-    errors.add(:pledge_sent, 'Clinic name cannot be blank') if clinic.blank?
-    errors.add(:pledge_sent, 'Appointment date cannot be blank') if appointment_date.blank?
   end
 end
